requirements.txt
streamlit
google-generativeai
Pillow
python-dotenv
